refactor: Phase 2 — split formatparse-core types/regex module
Move free functions and constants into regex/helpers.rs, tests into
regex/tests.rs, and keep FieldSpec::to_regex_pattern in regex/mod.rs.
Preserves the public strftime_to_regex export and all regex unit tests.
