+++ formatparse-core/src/types/regex/mod.rs
@@ -1,105 +1,15 @@
+//! Regex generation from field specifications.
+
+mod helpers;
+#[cfg(test)]
+mod tests;
+
+pub use helpers::strftime_to_regex;
+
 use crate::types::definitions::{FieldSpec, FieldType};
+use helpers::{int_width_precision_bounds, wrap_field_lookbehind, STRING_WIDTH_PRECISION_CHAR};
 use regex;
 use std::collections::HashMap;
-
-/// When [`FieldSpec::width`] and [`FieldSpec::precision`] are both set on integer fields
-/// (`d` / `i` / `x` / `X` / `o` / `b`), the significant digit run uses a bounded repetition
-/// `{min_digits,max_digits}` (inclusive).
-///
-/// This follows the **parse** library documentation (`parse#107`): width is a **minimum**
-/// number of digits and precision a **maximum**. Python’s `str.format` rejects `.precision`
-/// on integer `d` presentation types; formatparse still accepts it in patterns and applies
-/// these bounds so adjacent numeric fields do not greedily consume more digits than the
-/// maximum. The `next_field_is_greedy` hint used for width-only **string** fields does not
-/// apply here—the digit count is already capped by `precision`.
-///
-/// Returns [`None`] when the pair should fall back to legacy rules: missing width or
-/// precision, or **zero-padded** fields with `width > precision` (treated as width-only
-/// `{1,width}` so wide zero-filled fields are not collapsed to `precision` digits).
-///
-/// For non-zero-padded fields with `width > precision` (an inconsistent min/max pair), the
-/// quantifier degenerates to exactly `precision` digits.
-fn int_width_precision_bounds(
-    width: Option<usize>,
-    precision: Option<usize>,
-    zero_pad: bool,
-) -> Option<(usize, usize)> {
-    let w = width?;
-    let p = precision?;
-    if w <= p {
-        Some((w, p))
-    } else if zero_pad {
-        None
-    } else {
-        Some((p, p))
-    }
-}
-
-/// Convert strftime format string to regex pattern
-pub fn strftime_to_regex(format_str: &str) -> String {
-    let mut regex_parts = Vec::new();
-    let mut chars = format_str.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        if ch == '%' {
-            if let Some(next_ch) = chars.next() {
-                let regex_part = match next_ch {
-                    'Y' => r"\d{4}",             // Year with century
-                    'y' => r"\d{2}",             // Year without century
-                    'm' => r"\d{1,2}",           // Month (1-12 or 01-12) - flexible
-                    'd' => r"\d{1,2}",           // Day (1-31 or 01-31) - flexible
-                    'H' => r"\d{1,2}",           // Hour (0-23 or 00-23) - flexible
-                    'M' => r"\d{1,2}",           // Minute (0-59 or 00-59) - flexible
-                    'S' => r"\d{1,2}",           // Second (0-59 or 00-59) - flexible
-                    'b' | 'h' => r"[A-Za-z]{3}", // Abbreviated month name
-                    'B' => r"[A-Za-z]+",         // Full month name
-                    'a' => r"[A-Za-z]{3}",       // Abbreviated weekday
-                    'A' => r"[A-Za-z]+",         // Full weekday
-                    'w' => r"\d",                // Weekday as decimal (0=Sunday)
-                    'j' => r"\d{1,3}",           // Day of year (1-366, flexible padding)
-                    'U' | 'W' => r"\d{2}",       // Week number
-                    'c' => r".+",                // Date and time representation (locale dependent)
-                    'x' => r".+",                // Date representation (locale dependent)
-                    'X' => r".+",                // Time representation (locale dependent)
-                    '%' => "%",                  // Literal %
-                    _ => ".+?",                  // Unknown directive - match anything
-                };
-                regex_parts.push(regex_part.to_string());
-            }
-        } else {
-            // Escape special regex characters for literal text
-            regex_parts.push(regex::escape(&ch.to_string()));
-        }
-    }
-
-    regex_parts.join("")
-}
-
-/// Prefix the field body with optional zero-width lookbehind (issue #9).
-/// Trailing lookahead is appended **outside** the named/numbered capture in the pattern
-/// builder so a full-string anchor (`$`) can still match asserted suffix text.
-#[inline]
-fn wrap_field_lookbehind(spec: &FieldSpec, core: String) -> String {
-    format!("{}{}", spec.regex_lookbehind.as_deref().unwrap_or(""), core)
-}
-
-/// Character class for **one unit** of width/precision content in normal string fields (`s` / `{}`).
-///
-/// These fragments are embedded in parse regexes compiled with `(?s)` (DOTALL), so a bare `.`
-/// would match newlines and could swallow a literal line break after the field (issue #95).
-///
-/// We exclude CR/LF plus other line/paragraph boundaries that `str.splitlines()` treats as line
-/// ends in common Unicode text: VT, FF, NEL (U+0085), LINE SEPARATOR (U+2028), PARAGRAPH
-/// SEPARATOR (U+2029). Multiline (`:ml`) / indented block (`:blk`) fields still use `[\s\S]`
-/// where newlines are intentional.
-const STRING_WIDTH_PRECISION_CHAR: &str = concat!(
-    "[^", "\r", "\n", "\u{000B}", // VT
-    "\u{000C}", // FF
-    "\u{0085}", // NEL (Next Line)
-    "\u{2028}", // LINE SEPARATOR
-    "\u{2029}", // PARAGRAPH SEPARATOR
-    "]"
-);
 
 impl FieldSpec {
     /// `allow_empty_delimited`: for default unconstrained string fields only (no width,
@@ -540,497 +450,3 @@ impl FieldSpec {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::definitions::{FieldSpec, FieldType};
-    use fancy_regex::Regex;
-
-    #[test]
-    fn test_strftime_to_regex_year() {
-        assert_eq!(strftime_to_regex("%Y"), r"\d{4}");
-        assert_eq!(strftime_to_regex("%y"), r"\d{2}");
-    }
-
-    #[test]
-    fn test_strftime_to_regex_date() {
-        assert_eq!(strftime_to_regex("%m"), r"\d{1,2}");
-        assert_eq!(strftime_to_regex("%d"), r"\d{1,2}");
-    }
-
-    #[test]
-    fn test_strftime_to_regex_month_names() {
-        assert_eq!(strftime_to_regex("%b"), r"[A-Za-z]{3}");
-        assert_eq!(strftime_to_regex("%B"), r"[A-Za-z]+");
-        assert_eq!(strftime_to_regex("%h"), r"[A-Za-z]{3}");
-    }
-
-    #[test]
-    fn test_strftime_to_regex_weekday() {
-        assert_eq!(strftime_to_regex("%a"), r"[A-Za-z]{3}");
-        assert_eq!(strftime_to_regex("%A"), r"[A-Za-z]+");
-        assert_eq!(strftime_to_regex("%w"), r"\d");
-    }
-
-    #[test]
-    fn test_strftime_to_regex_literal() {
-        assert_eq!(strftime_to_regex("%%"), "%");
-    }
-
-    #[test]
-    fn test_strftime_to_regex_complex() {
-        let result = strftime_to_regex("%Y-%m-%d");
-        assert!(result.contains(r"\d{4}"));
-        assert!(result.contains(r"\d{1,2}"));
-        // Should escape the dashes
-        assert!(result.contains(r"\-"));
-    }
-
-    #[test]
-    fn test_strftime_to_regex_unknown() {
-        let result = strftime_to_regex("%Z");
-        assert_eq!(result, ".+?");
-    }
-
-    #[test]
-    fn test_field_spec_indent_block_matches_multiline_regex() {
-        let mut ml = FieldSpec::new();
-        ml.field_type = FieldType::Multiline;
-        ml.width = Some(3);
-        let mut blk = FieldSpec::new();
-        blk.field_type = FieldType::IndentBlock;
-        blk.width = Some(3);
-        let custom = HashMap::new();
-        assert_eq!(
-            ml.to_regex_pattern(&custom, None, false),
-            blk.to_regex_pattern(&custom, None, false)
-        );
-    }
-
-    #[test]
-    fn test_field_spec_string_default() {
-        let spec = FieldSpec::new();
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r".+?");
-    }
-
-    #[test]
-    fn test_field_spec_string_delimited_allows_empty_capture() {
-        let spec = FieldSpec::new();
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, true);
-        assert_eq!(pattern, r".*?");
-    }
-
-    #[test]
-    fn test_field_spec_multiline_with_width() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Multiline;
-        spec.width = Some(3);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r"[\s\S]{3,}");
-    }
-
-    #[test]
-    fn test_field_spec_multiline_with_width_exact_next_field() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Multiline;
-        spec.width = Some(4);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), Some(false), false);
-        assert_eq!(pattern, r"[\s\S]{4}");
-    }
-
-    #[test]
-    fn test_field_spec_multiline_align_right() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Multiline;
-        spec.alignment = Some('>');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r" *([\s\S]+?)");
-    }
-
-    #[test]
-    fn test_field_spec_multiline_precision_left_fill() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Multiline;
-        spec.precision = Some(3);
-        spec.alignment = Some('<');
-        spec.fill = Some('.');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r"[\s\S]{3}(?:\.*)");
-    }
-
-    #[test]
-    fn test_field_spec_string_with_precision() {
-        let mut spec = FieldSpec::new();
-        spec.precision = Some(5);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{5}");
-        assert_eq!(pattern, expected);
-    }
-
-    #[test]
-    fn test_field_spec_string_with_width() {
-        let mut spec = FieldSpec::new();
-        spec.width = Some(10);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{10,}");
-        assert_eq!(pattern, expected);
-    }
-
-    #[test]
-    fn test_field_spec_string_with_width_next_greedy() {
-        let mut spec = FieldSpec::new();
-        spec.width = Some(10);
-        // When next field is greedy, use greedy pattern
-        let pattern = spec.to_regex_pattern(&HashMap::new(), Some(true), false);
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{10,}");
-        assert_eq!(pattern, expected);
-    }
-
-    #[test]
-    fn test_field_spec_string_with_width_next_non_greedy() {
-        let mut spec = FieldSpec::new();
-        spec.width = Some(10);
-        // When next field is non-greedy (like {}), use exact width
-        let pattern = spec.to_regex_pattern(&HashMap::new(), Some(false), false);
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{10}");
-        assert_eq!(pattern, expected);
-    }
-
-    #[test]
-    fn test_string_precision_under_dotall_rejects_line_separator_in_run() {
-        use regex::Regex;
-        let mut spec = FieldSpec::new();
-        spec.precision = Some(4);
-        let p = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let re = Regex::new(&format!(r"(?s)^{}$", p)).unwrap();
-        assert!(re.is_match("abcd"));
-        assert!(!re.is_match("abc\u{2028}"));
-        assert!(!re.is_match("abc\u{2029}"));
-        assert!(!re.is_match("abc\u{0085}"));
-        assert!(!re.is_match("ab\u{000B}c"));
-    }
-
-    #[test]
-    fn test_field_spec_string_with_alignment_left() {
-        let mut spec = FieldSpec::new();
-        spec.alignment = Some('<');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"([^\{\}\s]+"));
-    }
-
-    #[test]
-    fn test_field_spec_string_with_alignment_right() {
-        let mut spec = FieldSpec::new();
-        spec.alignment = Some('>');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r" *(.+?)");
-    }
-
-    #[test]
-    fn test_field_spec_string_with_alignment_center() {
-        let mut spec = FieldSpec::new();
-        spec.alignment = Some('^');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"([^\{\}\s]+"));
-    }
-
-    #[test]
-    fn test_field_spec_string_with_precision_and_alignment() {
-        let mut spec = FieldSpec::new();
-        spec.precision = Some(5);
-        spec.alignment = Some('<');
-        spec.fill = Some('x');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{5}");
-        assert!(pattern.contains(&expected));
-        assert!(pattern.contains("x*"));
-    }
-
-    #[test]
-    fn test_field_spec_string_width_eq_precision_omits_optional_fill() {
-        // Issue #97: width == precision is an exact-width cell; optional fill after `{prec}`
-        // would greedily consume characters that belong to a following literal.
-        let mut spec = FieldSpec::new();
-        spec.precision = Some(5);
-        spec.width = Some(5);
-        spec.alignment = Some('<');
-        let mut expected = super::STRING_WIDTH_PRECISION_CHAR.to_string();
-        expected.push_str("{5}");
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, expected);
-
-        let mut spec_r = FieldSpec::new();
-        spec_r.precision = Some(5);
-        spec_r.width = Some(5);
-        spec_r.alignment = Some('>');
-        assert_eq!(
-            spec_r.to_regex_pattern(&HashMap::new(), None, false),
-            expected
-        );
-
-        let mut spec_c = FieldSpec::new();
-        spec_c.precision = Some(5);
-        spec_c.width = Some(5);
-        spec_c.alignment = Some('^');
-        assert_eq!(
-            spec_c.to_regex_pattern(&HashMap::new(), None, false),
-            expected
-        );
-    }
-
-    #[test]
-    fn test_field_spec_integer() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[+-]?"));
-        assert!(pattern.contains(r"\s*[0-9]+") || pattern.contains(r"[0-9]+"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_decimal_leading_whitespace_before_digits() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        let p = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let re = Regex::new(&format!("^{}$", p)).unwrap();
-        assert!(re.is_match("    0").unwrap());
-        assert!(re.is_match("  42").unwrap());
-    }
-
-    #[test]
-    fn test_field_spec_integer_with_zero_pad() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.zero_pad = true;
-        spec.width = Some(5);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("[0-9]{1,5}"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_hex() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.original_type_char = Some('x');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("0[xX]"));
-        assert!(pattern.contains("[0-9a-fA-F]+"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_octal() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.original_type_char = Some('o');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("0[oO]"));
-        assert!(pattern.contains("[0-7]+"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_binary() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.original_type_char = Some('b');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("0[bB]"));
-        assert!(pattern.contains("[01]+"));
-    }
-
-    #[test]
-    fn test_field_spec_float() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Float;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[+-]?"));
-        assert!(pattern.contains(r"\d+\.\d+"));
-    }
-
-    #[test]
-    fn test_field_spec_float_with_precision() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Float;
-        spec.precision = Some(2);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"\.\d{2}"));
-    }
-
-    #[test]
-    fn test_field_spec_float_precision_zero_accepts_integer_like_text() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Float;
-        spec.precision = Some(0);
-        spec.width = Some(2);
-        let p = spec.to_regex_pattern(&HashMap::new(), None, false);
-        let re = Regex::new(&format!("^{}$", p)).unwrap();
-        assert!(re.is_match("20").unwrap());
-        assert!(re.is_match(" 20").unwrap());
-        assert!(re.is_match("20.000").unwrap());
-    }
-
-    #[test]
-    fn test_field_spec_boolean() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Boolean;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("true"));
-        assert!(pattern.contains("false"));
-        assert!(pattern.contains("1"));
-        assert!(pattern.contains("0"));
-    }
-
-    #[test]
-    fn test_field_spec_letters() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Letters;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r"[a-zA-Z]+");
-    }
-
-    #[test]
-    fn test_field_spec_word() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Word;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r"\w+");
-    }
-
-    #[test]
-    fn test_field_spec_datetime_iso() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::DateTimeISO;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"\d{4}-\d{2}-\d{2}"));
-    }
-
-    #[test]
-    fn test_field_spec_datetime_strftime() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::DateTimeStrftime;
-        spec.strftime_format = Some("%Y-%m-%d".to_string());
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"\d{4}"));
-        assert!(pattern.contains(r"\d{1,2}"));
-    }
-
-    #[test]
-    fn test_field_spec_custom_type() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Custom("MyType".to_string());
-        let mut custom_patterns = HashMap::new();
-        custom_patterns.insert("MyType".to_string(), r"\d+".to_string());
-        let pattern = spec.to_regex_pattern(&custom_patterns, None, false);
-        assert_eq!(pattern, r"\d+");
-    }
-
-    #[test]
-    fn test_field_spec_custom_type_no_pattern() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Custom("MyType".to_string());
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        // Should default to non-whitespace
-        assert_eq!(pattern, r"\S+");
-    }
-
-    #[test]
-    fn test_field_spec_braced_content_placeholder_regex() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::BracedContent;
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert_eq!(pattern, r".*?");
-    }
-
-    #[test]
-    fn test_field_spec_integer_sign_plus() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.sign = Some('+');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"\+?"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_sign_space() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.sign = Some(' ');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[- ]?"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_fill_equals_alignment() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.fill = Some('x');
-        spec.alignment = Some('=');
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        // Should have fill pattern between sign and digits
-        assert!(pattern.contains("x*"));
-    }
-
-    #[test]
-    fn test_field_spec_integer_decimal_width_and_precision_bounded() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.width = Some(2);
-        spec.precision = Some(2);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[0-9]{2,2}"), "pattern: {}", pattern);
-        let re = Regex::new(&format!("^{}$", pattern)).unwrap();
-        assert!(re.is_match("99").unwrap());
-        assert!(!re.is_match("999").unwrap());
-    }
-
-    #[test]
-    fn test_field_spec_integer_hex_width_and_precision_bounded() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.original_type_char = Some('x');
-        spec.width = Some(2);
-        spec.precision = Some(2);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(
-            pattern.contains(r"[0-9a-fA-F]{2,2}"),
-            "pattern: {}",
-            pattern
-        );
-    }
-
-    #[test]
-    fn test_field_spec_integer_width_gt_precision_nonzero_pad_degenerates() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.width = Some(5);
-        spec.precision = Some(2);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[0-9]{2,2}"), "pattern: {}", pattern);
-    }
-
-    #[test]
-    fn test_field_spec_integer_width_gt_precision_zero_pad_uses_width_only() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.zero_pad = true;
-        spec.width = Some(30);
-        spec.precision = Some(2);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains("[0-9]{1,30}"), "pattern: {}", pattern);
-    }
-
-    #[test]
-    fn test_field_spec_integer_decimal_width_2_precision_5() {
-        let mut spec = FieldSpec::new();
-        spec.field_type = FieldType::Integer;
-        spec.width = Some(2);
-        spec.precision = Some(5);
-        let pattern = spec.to_regex_pattern(&HashMap::new(), None, false);
-        assert!(pattern.contains(r"[0-9]{2,5}"), "pattern: {}", pattern);
-    }
-}
